refactored ace.system.base ns
